Infra/requirements.txt
apache-airflow==2.6.3
apache-airflow-providers-apache-livy==3.5.2
apache-airflow-providers-apache-spark==4.1.1
apache-airflow-providers-cncf-kubernetes==7.6.0
apache-airflow-providers-common-sql==1.6.0
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-http==4.5.0
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-sqlite==3.4.2
apache-airflow-providers-ssh==3.7.3
avro==1.11.2
avro-python3==1.10.2
gunicorn==21.2.0
paramiko==3.3.1
pyarrow==12.0.1
pyspark==3.4.0
PyYAML==5.4.1
redis==4.6.0
mysql-connector-python
celery
influxdb-client
kafka-python